fix: move halving logic to BriocheConfig
